organicé las carpetas de redux y corregí algunos errores que encontré
