+++ client/src/redux/store/index.js
@@ -1,11 +1,12 @@
 import { applyMiddleware, createStore } from "redux";
 import { rootReducer } from "../reducer";
-import { composeWithDevTooLs } from "redux-devtools-extension";
-import { thunk } from "redux-thunk";
+import { composeWithDevTools } from "redux-devtools-extension";
+import thunk from "redux-thunk";
+
 
 const store = createStore(
     rootReducer,
-    composeWithDevTooLs(applyMiddleware(thunk))
+    composeWithDevTools(applyMiddleware(thunk))
 
 );
